requirements.txt
openai
python-dotenv
aiohttp
asyncio
beautifulsoup4
lxml
pydantic
openai-agents
md2pdf
faiss-cpu>=1.7.4